client/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
geojson==3.1.0
shapely==2.0.1
simplekml==1.3.6
reportlab==4.0.8
openpyxl==3.1.2
Pillow==10.1.0
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
